Split and refactor prison config integration tests
